+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## 1.0.1
+
+- 对于 CarrierWave 的 cache 机制正确支持；
+- 对 UpYun 偶发的 `concurrent put or delete` 错误容错；
+- 清理不必要的 dependency.
+
+## 1.0.0
+
+- 支持 CarrierWave 2.0.0+;
+
 ## 0.2.2
 
 - 减少 Gem 文件尺寸，避免带入不必要的问题；